cleaning, refactor of task classes
